requirements_test.txt
fastapi
requests
azure-core
azure-storage-file-datalake
pandas
numpy
json
pydantic
asyncio
aiohttp
starlette_exporter
osiris-sdk
prometheus_client
jaeger_client


